requirements.txt
click==6.7
Django==1.11
djangorestframework==3.7.3
Jinja2==2.10
livereload==2.5.1
Markdown==2.6.10
MarkupSafe==1.0
mkdocs==0.17.2
pytz==2017.3
PyYAML==3.12
six==1.11.0
tornado==4.5.2
